BIG FIX - get rid of inc to define unique name
An entire WordPress admin page had to be filled up with AddBySearch
instances before this bug was found. It turns out the ‘inc’ wasn’t
working the way it should have. The solution was beneath my finger tips
all along. To find a unique id, all I had to do was use the field name.
Duh!